inw is a 16 inw는 16입니다.

+++ libc/stdio/stdio.h
@@ -14,7 +14,7 @@ void outb(uint16 port, uint8 data)
   asm volatile("outb %0, %1" : : "a"(data), "Nd"(port));
 }
 
-uint8 inw(uint16 port)
+uint16 inw(uint16 port)
 {
   uint16 data;
   asm volatile("inw %1, %0" : "=a"(data) : "Nd"(port));
